backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pymongo==4.6.1
python-dotenv==1.0.1
sentence-transformers==2.5.1
google-generativeai==0.3.2
python-multipart==0.0.9
pydantic==2.6.1 